+++ plone/app/contenttypes/migration/utils.py
@@ -213,6 +213,12 @@ def migrate_leadimage(source_object, target_object):
                     "Could not migrate collective.leadimage fields.")
         return
 
+    acc = source_object.getField(
+        OLD_LEADIMAGE_FIELD_NAME).getAccessor(source_object)()
+    if getattr(acc, 'filename', None) is None:
+        # skip if old content has field but has no lead image in the field
+        return
+
     # handle image field
     migrate_imagefield(
         source_object,
@@ -226,8 +232,8 @@ def migrate_leadimage(source_object, target_object):
         target_object,
         OLD_CAPTION_FIELD_NAME,
         NEW_CAPTION_FIELD_NAME)
-    logger.info(
-        "Migrating contentlead image for." % target_object.absolute_url())
+    logger.info("Migrating contentlead image for {0}.".format(
+        target_object.absolute_url()))
 
 
 def migrate_portlets(src_obj, dst_obj):